JBJCA-863 Add support for maximum-free-thread scheduler in DWM

+++ core/src/main/java/org/jboss/jca/core/spi/workmanager/notification/NotificationListener.java
@@ -44,12 +44,12 @@ public interface NotificationListener
     * @param id The work manager id
     * @param free The number of free threads
     */
-   public void updateShortRunningFree(String id, int free);
+   public void updateShortRunningFree(String id, long free);
 
    /**
     * Update the long thread pool information for a distributed work manager
     * @param id The work manager id
     * @param free The number of free threads
     */
-   public void updateLongRunningFree(String id, int free);
+   public void updateLongRunningFree(String id, long free);
 }
