+++ core/src/main/java/org/jboss/jca/core/workmanager/selector/AbstractSelector.java
@@ -47,10 +47,10 @@ public abstract class AbstractSelector implements Selector, NotificationListener
    protected DistributedWorkManager dwm;
 
    /** Short running */
-   protected Map<String, Integer> shortRunning;
+   protected Map<String, Long> shortRunning;
 
    /** Long running */
-   protected Map<String, Integer> longRunning;
+   protected Map<String, Long> longRunning;
 
    /**
     * Constructor
@@ -58,8 +58,8 @@ public abstract class AbstractSelector implements Selector, NotificationListener
    public AbstractSelector()
    {
       this.dwm = null;
-      this.shortRunning = Collections.synchronizedMap(new HashMap<String, Integer>());
-      this.longRunning = Collections.synchronizedMap(new HashMap<String, Integer>());
+      this.shortRunning = Collections.synchronizedMap(new HashMap<String, Long>());
+      this.longRunning = Collections.synchronizedMap(new HashMap<String, Long>());
    }
 
    /**
@@ -91,17 +91,17 @@ public abstract class AbstractSelector implements Selector, NotificationListener
    /**
     * {@inheritDoc}
     */
-   public void updateShortRunningFree(String id, int free)
+   public void updateShortRunningFree(String id, long free)
    {
-      shortRunning.put(id, Integer.valueOf(free));
+      shortRunning.put(id, Long.valueOf(free));
    }
 
    /**
     * {@inheritDoc}
     */
-   public void updateLongRunningFree(String id, int free)
+   public void updateLongRunningFree(String id, long free)
    {
-      longRunning.put(id, Integer.valueOf(free));
+      longRunning.put(id, Long.valueOf(free));
    }
 
    /**
@@ -147,16 +147,16 @@ public abstract class AbstractSelector implements Selector, NotificationListener
     * @param work The work instance
     * @return The selection map
     */
-   protected Map<String, Integer> getSelectionMap(DistributableWork work)
+   protected Map<String, Long> getSelectionMap(DistributableWork work)
    {
-      Map<String, Integer> sorted = null;
+      Map<String, Long> sorted = null;
       if (isLongRunning(work))
       {
-         sorted = new HashMap<String, Integer>(longRunning);
+         sorted = new HashMap<String, Long>(longRunning);
       }
       else
       {
-         sorted = new HashMap<String, Integer>(shortRunning);
+         sorted = new HashMap<String, Long>(shortRunning);
       }
 
       return sorted;
